requirements.txt
asgiref==3.11.0
Django==6.0.1
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==10.4.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
